unifica cadastro e ajusta controle de acesso por role
- Unificado endpoint de cadastro de usuário (admin e autocadastro em um só)
- Implementada lógica para diferenciar criação via admin ou usuário comum
- Ajustado service para definir role/tipo dinamicamente com base no usuário logado
- Adicionado suporte a @AuthenticationPrincipal no controller
- Corrigidas validações e tratamento de dados (CPF, telefone, idade, etc.)
- Implementado controle de acesso: ADMIN pode tudo, USER apenas a si mesmo

+++ src/main/java/com/wictor/Security/JwtFilter.java
@@ -44,10 +44,11 @@ public class JwtFilter extends OncePerRequestFilter {
                 String token = authHeader.replace("Bearer ", "");
 
                 if (jwtService.validarToken(token)) {
-                    Long userId = jwtService.extrairId(token);
+                    Integer userId = jwtService.extrairId(token);
                     User user = userRepository.findById(userId).orElse(null);
 
                     if (user != null && user.isAtivo()) {
+
                         UsernamePasswordAuthenticationToken auth =
                                 new UsernamePasswordAuthenticationToken(
                                         user,
